# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TeklifPro API",
    "version": "1.0.0",
    "description": "TeklifPro is a Turkish B2B proposal management platform designed to streamline the creation, management, and delivery of business proposals. The API provides comprehensive endpoints for authentication, proposal management, customer management, product catalogs, AI-powered suggestions, and integration with accounting systems via Parasut.",
    "contact": {
      "name": "Sercan Hosgoerns",
      "email": "hosgorens@gmail.com"
    },
    "license": {
      "name": "All Rights Reserved"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.teklifpro.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/proposals/{id}/send": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "summary": "Send proposal",
        "description": "Sends a proposal to the customer via WhatsApp or Email. | Teklifi müşteriye WhatsApp veya E-posta yoluyla gönderir.",
        "operationId": "sendProposal",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "channel"
                ],
                "properties": {
                  "channel": {
                    "type": "string",
                    "enum": [
                      "EMAIL",
                      "WHATSAPP"
                    ],
                    "example": "WHATSAPP"
                  },
                  "message": {
                    "type": "string",
                    "example": "Merhaba, aşağıda size hazırlanmış teklifimizi bulabilirsiniz."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proposal sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "sentAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "channel": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid channel or proposal cannot be sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Proposal not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/proposals/bulk-send": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "summary": "Bulk send proposals",
        "description": "Sends multiple proposals to different customers in a single request. | Birden fazla teklifi farklı müşterilere tek bir istekte gönderir.",
        "operationId": "bulkSendProposals",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "proposals"
                ],
                "properties": {
                  "proposals": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": [
                        "proposalId",
                        "channel"
                      ],
                      "properties": {
                        "proposalId": {
                          "type": "string"
                        },
                        "channel": {
                          "type": "string",
                          "enum": [
                            "EMAIL",
                            "WHATSAPP"
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proposals sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "successful": {
                          "type": "integer"
                        },
                        "failed": {
                          "type": "integer"
                        },
                        "results": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "proposalId": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string",
                                "enum": [
                                  "success",
                                  "failed"
                                ]
                              },
                              "error": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from registration or login"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input parameters"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
    }
  }
}